EPROD-1082 migrated dfx tests to pocket-ic
