+++ Cargo.toml
@@ -32,36 +32,36 @@ bitcoincore-rpc = "0.18.0"
 candid = { version = "0.10", features = ["value"] }
 cfg-if = "1.0"
 clap = { version = "4", features = ["derive"] }
-did = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "did", tag = "v0.34.x", features = [
+did = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "did", branch = "EPROD-1082-temp-deps", features = [
   "alloy-primitives-07",
 ] }
 dirs = "5"
 dotenv = "0.15"
 drop_guard = "0.3"
 env_logger = { version = "0.11.4", default-features = false }
-eth-signer = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "eth-signer", tag = "v0.34.x" }
-ethereum-json-rpc-client = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "ethereum-json-rpc-client", tag = "v0.34.x" }
+eth-signer = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "eth-signer", branch = "EPROD-1082-temp-deps" }
+ethereum-json-rpc-client = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "ethereum-json-rpc-client", branch = "EPROD-1082-temp-deps" }
 ethereum-types = "0.14"
 ethers-core = "2.0"
-evm-canister-client = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "evm-canister-client", tag = "v0.34.x" }
+evm-canister-client = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "evm-canister-client", branch = "EPROD-1082-temp-deps" }
 foundry-compilers = { version = "0.9", features = ["svm-solc"] }
 futures = { version = "0.3", default-features = false }
 hex = "0.4"
-ic-agent = "0.37"
+ic-agent = "0.39"
 ic-btc-interface = "0.2"
-ic-canister = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-canister", tag = "v0.22.x" }
-ic-canister-client = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-canister-client", tag = "v0.22.x" }
-ic-exports = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-exports", tag = "v0.22.x" }
-ic-log = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-log", tag = "v0.22.x" }
-ic-metrics = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-metrics", tag = "v0.22.x", features = [
+ic-canister = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-canister", branch = "EPROD-1082-migrate-all-dfx-and-state-machine-tests-to-pocket-ic" }
+ic-canister-client = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-canister-client", branch = "EPROD-1082-migrate-all-dfx-and-state-machine-tests-to-pocket-ic" }
+ic-exports = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-exports", branch = "EPROD-1082-migrate-all-dfx-and-state-machine-tests-to-pocket-ic" }
+ic-log = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-log", branch = "EPROD-1082-migrate-all-dfx-and-state-machine-tests-to-pocket-ic" }
+ic-metrics = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-metrics", branch = "EPROD-1082-migrate-all-dfx-and-state-machine-tests-to-pocket-ic", features = [
   "export-api",
 ] }
-ic-stable-structures = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-stable-structures", tag = "v0.22.x" }
-ic-storage = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-storage", tag = "v0.22.x" }
-ic-task-scheduler = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-task-scheduler", tag = "v0.22.x" }
-ic-test-utils = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-test-utils", tag = "v0.22.x" }
-ic-utils = "0.37"
-icrc-client = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "icrc-client", tag = "v0.34.x" }
+ic-stable-structures = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-stable-structures", branch = "EPROD-1082-migrate-all-dfx-and-state-machine-tests-to-pocket-ic" }
+ic-storage = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-storage", branch = "EPROD-1082-migrate-all-dfx-and-state-machine-tests-to-pocket-ic" }
+ic-task-scheduler = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-task-scheduler", branch = "EPROD-1082-migrate-all-dfx-and-state-machine-tests-to-pocket-ic" }
+ic-test-utils = { git = "https://github.com/bitfinity-network/canister-sdk", package = "ic-test-utils", branch = "EPROD-1082-migrate-all-dfx-and-state-machine-tests-to-pocket-ic" }
+ic-utils = "0.39"
+icrc-client = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "icrc-client", branch = "EPROD-1082-temp-deps" }
 jsonrpc-core = "18.0"
 lazy-regex = "3"
 log = "0.4"
@@ -78,7 +78,7 @@ serde = "1.0"
 serde_bytes = "0.11"
 serde_json = "1.0"
 serial_test = "3"
-signature-verification-canister-client = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "signature-verification-canister-client", tag = "v0.34.x" }
+signature-verification-canister-client = { git = "https://github.com/bitfinity-network/bitfinity-evm-sdk", package = "signature-verification-canister-client", branch = "EPROD-1082-temp-deps" }
 snapbox = "0.6"
 strum = "0.26"
 tempfile = "3"
