+++ just/test.just
@@ -1,3 +1,4 @@
+TEST_LOGFILE:="./target/integration_tests.log"
 
 # Run all unit tests and integration tests
 [group('test')]
@@ -12,9 +13,9 @@ test test_name="":
 
 # Run pocket-ic and state machine integration tests
 [group('test')]
-integration_test test_name="": 
+integration_test test_name="": docker_down docker_pull docker_up
   export PROTOC_INCLUDE=${PWD}/proto
-  cargo test {{test_name}} -p integration-tests --no-default-features --features "pocket_ic_integration_test" --features "state_machine_tests"
+  cargo test {{test_name}} -p integration-tests --no-default-features --features "pocket_ic_integration_test" --features "state_machine_tests" 2> "{{TEST_LOGFILE}}"
 
 
 # Run all unit tests for the i686 target
