+++ src/integration-tests/tests/utils/wasm.rs
@@ -5,13 +5,13 @@ use tokio::io::AsyncReadExt;
 
 use crate::utils::get_workspace_root_dir;
 
-const BTC_CANISTER_MOCK_WASM_FILENAME: &str = "ic-bitcoin-canister-mock.wasm.gz";
 const IC_CKBTC_MINTER_WASM_FILENAME: &str = "ic-ckbtc-minter.wasm.gz";
 const IC_CKBTC_KYT_WASM_FILENAME: &str = "ic-ckbtc-kyt.wasm.gz";
 const EVM_RPC_WASM_FILENAME: &str = "evm_rpc.wasm.gz";
 const BTC_BRIDGE_WASM_FILENAME: &str = "btc-bridge.wasm.gz";
 const ERC20_BRIDGE_WASM_FILENAME: &str = "erc20-bridge.wasm.gz";
-const BTC_CANISTER_WASM_FILENAME: &str = "ic-bitcoin-canister-mock.wasm.gz";
+const BTC_MOCK_CANISTER_WASM_FILENAME: &str = "ic-bitcoin-canister-mock.wasm.gz";
+const BTC_CANISTER_WASM_FILENAME: &str = "ic-btc-canister.wasm.gz";
 const SIGNATURE_VERIFICATION_WASM_FILENAME: &str = "signature_verification.wasm.gz";
 const EVM_TESTNET_WASM_FILENAME: &str = "evm_testnet.wasm.gz";
 const ICRC1_LEDGER_WASM_FILENAME: &str = "icrc1-ledger.wasm.gz";
@@ -94,13 +94,23 @@ pub async fn get_icrc2_bridge_canister_wasm_path() -> PathBuf {
     get_wasm_path(ICRC2_BRIDGE_WASM_FILENAME).await
 }
 
-pub async fn get_btc_canister_bytecode() -> Vec<u8> {
+pub async fn get_mock_btc_canister_bytecode() -> Vec<u8> {
     static CANISTER_BYTECODE: OnceCell<Vec<u8>> = OnceCell::new();
-    get_or_load_wasm(&CANISTER_BYTECODE, BTC_CANISTER_MOCK_WASM_FILENAME).await
+    get_or_load_wasm(&CANISTER_BYTECODE, BTC_MOCK_CANISTER_WASM_FILENAME).await
 }
 
-/// Returns the path to the btc canister
-pub async fn get_btc_canister_wasm_path() -> PathBuf {
+/// Returns the path to the mocked btc canister
+pub async fn get_mock_btc_canister_wasm_path() -> PathBuf {
+    get_wasm_path(BTC_MOCK_CANISTER_WASM_FILENAME).await
+}
+
+pub async fn get_ic_btc_canister_bytecode() -> Vec<u8> {
+    static CANISTER_BYTECODE: OnceCell<Vec<u8>> = OnceCell::new();
+    get_or_load_wasm(&CANISTER_BYTECODE, BTC_CANISTER_WASM_FILENAME).await
+}
+
+/// Returns the path to the IC btc canister
+pub async fn get_ic_btc_canister_wasm_path() -> PathBuf {
     get_wasm_path(BTC_CANISTER_WASM_FILENAME).await
 }
